refactor(common/Layout): adjust Layout
