+++ app/components/common/Logo/index.tsx
