+++ app/features/favorites/index.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import PetCards from '~/components/common/PetCards';
 import type { PetCard } from '~/components/common/PetCards/types';
-import { HeaderPortal } from '~/components/Layout/header';
+import { HeaderPortal } from '~/components/common/Layout/header';
 import { getMockPets, TITLE, SUBTITLE } from './constants';
 
 export default function Favorites() {
