+++ app/routes/favorites/index.tsx
@@ -1,10 +1,5 @@
-import Layout from '~/components/Layout';
 import Favorites from '~/features/favorites';
 
 export default function ThemeColor() {
-  return (
-    <Layout>
-      <Favorites />
-    </Layout>
-  );
+  return <Favorites />;
 }
