+++ app/routes/hospital/index.tsx
@@ -1,10 +1,5 @@
-import Layout from '~/components/Layout';
 import Hospital from '~/features/hospital';
 
 export default function ThemeColor() {
-  return (
-    <Layout>
-      <Hospital />
-    </Layout>
-  );
+  return <Hospital />;
 }
