+++ app/routes/theme/color.tsx
@@ -1,10 +1,5 @@
-import Layout from '~/components/Layout';
 import ThemeColorCat from '~/features/theme/color';
 
 export default function ThemeColor() {
-  return (
-    <Layout>
-      <ThemeColorCat />
-    </Layout>
-  );
+  return <ThemeColorCat />;
 }
